OrderTable와 Order와의 양방향 의존성 해결을 위한 인터페이스 도입
